pal: don't convert distance from feature, leave it up to caller to specify exact distance
git-svn-id: http://svn.osgeo.org/qgis/branches/symbology-ng-branch@11020 c8812cc2-4d05-0410-92ff-de0c093fc19c

+++ src/core/pal/feature.h
@@ -84,7 +84,7 @@ namespace pal
       char *uid;
       Layer *layer;
 
-      int distlabel;
+      double distlabel;
 
       GEOSGeometry *the_geom;
       int currentAccess;
